Code refinement
While testing the code, I found that ```c[i+j].coeff``` was getting reassigned different values as i and j were incrementing in the nested for loop of ```c[i + j].coeff += a[i].coeff * b[j].coeff;```, that is c[0+1] and c[1+0] was assigning the same coefficient to two completely different variables of the polynomial. A few changes have been made to the code to address this.  

- Instead of storing the coefficients and exponents in the ```c[i+j]``` index, I stored it in ```c[k]```.
- This gave me the correct results, however the variables were no longer grouped. Another loop was introduced in the code that targeted identical coefficients of the resultant polynomial and its corresponding coefficients were added.